Better rhel setup

+++ rhel.sh
@@ -17,8 +17,8 @@ cd -
 # chmod 400 /home/duybui/.ssh/id_rsa
 # git clone git@bitbucket.org:khuongduybui/linux.git /home/duybui/setup
 
-wget http://mirror.sfo12.us.leaseweb.net/epel/5/i386/epel-release-5-4.noarch.rpm
-sudo yum install -y http://mirror.sfo12.us.leaseweb.net/epel/5/i386/epel-release-5-4.noarch.rpm
+wget http://dl.fedoraproject.org/pub/archive/epel/5/i386//epel-release-5-4.noarch.rpm
+sudo yum install -y epel-release-5-4.noarch.rpm
 sudo yum check-update
 
 # wget ftp://ftp.pbone.net/mirror/download.fedora.redhat.com/pub/fedora/epel/5/x86_64/tmux-1.4-3.el5.1.x86_64.rpm
@@ -27,8 +27,10 @@ sudo yum check-update
 # yum install -y byobu-5.73-4.el5.noarch.rpm
 sudo yum install -y byobu
 # byobu-launcher-install
-mv /home/duybui/.zprofile /home/duybui/.zprofile.`date +%Y-%m-%d`.bak
+test -f /home/duybui/.zprofile && mv /home/duybui/.zprofile /home/duybui/.zprofile.`date +%Y-%m-%d`.bak
 echo "_byobu_sourced=1 byobu -S ~/duybui new-session -A -s duybui" > /home/duybui/.zprofile
+echo "set -g default-shell /usr/bin/fish" > $HOME/.byobu/.tmux.conf
+echo "set -g default-command /usr/bin/fish" >> $HOME/.byobu/.tmux.conf
 # exec byobu-launcher
 
 # yum install -y w3m
